AI::AutoFire: make enemies[] a vector of references to shared_ptr
The ships list holds references to all ships, and creating / destroying refs to
to many ships repeatedly is not needed.  (And is not cheap, because shared_ptr
uses thread-safe atomics.)

Better: make some sub-lists to save time for AI functions, like list of
candidate enemies for each government in-system, and pass those in to
per-ship functions like AutoFire().

---
TODO: separate commit for this: order the enemies checks in order of
cheapness and filtering power: GetSystem is a trivial pointer-compare
and can potentially filter out a lot of ships.  Position is pretty cheap
(given improvements to Point).  Government checks are still expensive
(map lookup).  IsTargetable() is a bunch of little things that are probably
rare.

IsTargetable includes Ship::forget, rather than Ship::isInSystem, probably because
isInSystem isn't correct in all ships until AI::Fire() has been called on them.

AI::AutoFire() is called while isInSystem is set correctly.

The forget check will reject ships that were out-of-system last frame, but
not ships in their first frame out of system.  It's easiest to just leave
the GetSystem() check in place.