Point: vectorize with SSE2.  Make it a non-aggregate POD type, so it's passed in XMM regs
This requires using ugly pointer casts to type pun between double and __m128d
if we want to avoid a union (to make the pass and return by value semantics
put the whole struct in a single register, rather than x and y in separate regs.

Apparently this is legal in GNU C (i.e. gcc, clang, and icc).
IDK about MSVC or other compilers, but the code is already not portable to them.
(Some use of GNU C vector extensions, like v * v instead of _mm_mul_pd(v,v).

remove  using namespace std from Point.h, but be lazy and add it in to Ship.h
This required using std:: on inlined stuff in Outfit.h