Mask::Contains(Point) tune the loop, and a main() with a microbenchmark
Gives a ~1.5x speedup for this function on Core2.

The prev.X() != next.X() is 100% redundant with the other check (except
for NaNs, where the answer is meaningless either way).

 p <  PX=NX   false == true:  PX!=NX is redundant
 p == PX=NX   true  == false: PX!=NX is redundant
 p >  PX=NX   true  == false: PX!=NX is redundant
 p unord PX=NX false == false: NaN makes the whole thing meaningless

It's a significant slowdown, not a speedup on Core2.  It's extra work
for a rare early out, and the out causes a branch misprediction.
Also, gcc5.2 happens to lay out the loop in a worse way for Core2.
Without it, there's a small loop that fits in Core2's 64B pre-decode buffer

C2Duo E6600 (2.4GHz) microbenchmark of 1000000 reps
 of 8 different points: (x=y = i*radius/8.)
  g++5.2 -O3 -ffast-math -march=core2

without NX==PX: kestrel 1337 ms.  ~2.5% branch mispredicts.  1.0 billion branches
with it: kestrel 2116ms, ~4.6% branch mispredicts.  1.5 billion branches