Point: unary operator-: more efficient without -ffast-math
Use the unary operator instead of expressing it as subtraction from 0.

Without -ffast-math, Point() - *this actually compiles to a subtract from a
zeroed register.

Point(-v) compiles to an XORPD (but unfortunately still not to just reversing
the operands to a previous subtract, which is what happens with -ffast-math
for either method).