WithinRange: Optimize for AVX
With AVX:
allows the loads inside the loop to be a memory operand for VADDPS,
decreasing instruction count (and fused-domain uop count.  Works with or
without -ffast-math.

Without -ffast-math:
Costs an extra XORPD instruction outside the loop, regardless of AVX or not

With -ffast-math:
operands to the previous SUBPS are reversed, so there's no cost.