untested/experimental/WIP Mask: do Contains and Intersect in one pass
There seems to be a huge amount of redundancy between them, with some
rearranging to expose it.  This may require changing Contains from
closed start, open end to vice versa, so we can just use the sign bits.

Otherwise we still need a bunch of compares against 0.0, or compares as
well as subtracts.

This is untested, but if my optimistic guess about what might work is
correct, the asm compiles to only a few more instructions than Intersects.

Contains is only needed 1-2% of the time when Intersect is called, though,
so probably best not to spend too long on this.